+++ src/assets/misc/CHANGELOG.md
@@ -142,3 +142,5 @@
 - Reworked sound engine
 
 - Changelog now appears based on a hash (MD5) rather than the version number
+
+- Allowed mobile devices to be used, but with a warning